backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
python-dotenv==1.0.1
celery==5.6.2
django-celery-beat==2.9.0
sqlparse==0.5.5
tzdata==2025.3
redis==5.2.1
PyJWT==2.10.1
Pillow==11.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pytesseract==0.3.13
django-storages[boto3]==1.14.4
boto3==1.35.99
